Now bot also reads "Locations" sent by gps
Added a way to the bot to read "Locations" types of messages. Just
attach a location via telegram to the bot, and it will send the latitude
and longitude to the server (scanner).

I'm do not have a linux right now to test if the server will behave
correctly, but it should work.